refactor: moved typography + reset to foundations
