--- packages/aksara-ui-kit/src/components/index.ts
@@ -1,1 +1,0 @@
-export * from './reset';
