+++ packages/aksara-ui-kit/src/foundations/index.ts
@@ -1,0 +1,2 @@
+export * from './reset';
+export * from './typography';
